FIX: hack to fix issue when invoicing sale orders
Fix the following issue https://github.com/OCA/bank-payment/issues/536

The payment mode is overridden by Odoo.
There is no good fix to that, excepted to ask Odoo to change the way
the 'action_invoice_create()' method is implemented.

It has been stated with the PM that we have to fix it locally.